feat: feat set null for relations on delete artist
