+++ src/modules/artist/artist.service.ts
@@ -2,17 +2,11 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { ArtistEntity } from '../db/entities/artist/artist.entity';
-import { Album, Artist, Track } from 'src/models';
-import { AlbumEntity } from '../db/entities/album/album.entity';
-import { TrackEntity } from '../db/entities/track/track.entity';
+import { Artist } from 'src/models';
 
 @Injectable()
 export class ArtistService {
-  constructor(
-    private readonly artistEntity: ArtistEntity,
-    private readonly albumEntity: AlbumEntity,
-    private readonly trackEntity: TrackEntity,
-  ) {}
+  constructor(private readonly artistEntity: ArtistEntity) {}
 
   create(createArtistDto: CreateArtistDto): Artist {
     return this.artistEntity.create(createArtistDto);
@@ -44,25 +38,6 @@ export class ArtistService {
 
   remove(id: string) {
     this.findOne(id);
-
-    // TODO: refactor using Prisma
-    const artistAlbums: Album[] = this.albumEntity.findAllByArtistId(id);
-
-    artistAlbums.forEach((album) => {
-      this.albumEntity.update(album.id, {
-        ...album,
-        artistId: null,
-      });
-    });
-
-    const artistTracks: Track[] = this.trackEntity.findAllByArtistId(id);
-
-    artistTracks.forEach((track) => {
-      this.trackEntity.update(track.id, {
-        ...track,
-        artistId: null,
-      });
-    });
 
     this.artistEntity.remove(id);
   }
